fix plotting for small vals IK still gives negative angles for beta, gamma and wrong one for alpha
